feat: finish (kinda)  get_timeline endpoint

+++ src/domain/constants.rs
@@ -1,2 +1,2 @@
 pub const X_REQUEST_ID_HEADER_NAME: &str = "x-request-id";
-pub const TIMELINE_LIMIT: i8 = 20;
+pub const TIMELINE_LIMIT: i64 = 20;
